move all form related components into a form directory
a bit more code tidying. Move all the components related to forms into
its own directory.
