+++ frontend/src/form/components/SelectOption.tsx
@@ -1,4 +1,5 @@
 import { useComputed, type ReadonlySignal } from "@preact/signals";
+
 import { SelectOptionType } from "../types/SelectOptionType";
 
 export interface SelectOptionProps extends SelectOptionType {
