+++ frontend/src/form/components/TextInputRow.test.tsx
@@ -1,6 +1,6 @@
 import { afterEach, beforeEach, describe, expect, test, vi } from "vitest";
 
-import { renderWithProviders } from "../test/renderWithProviders";
+import { renderWithProviders } from "../../test/renderWithProviders";
 import { TextInputRow } from "./TextInputRow";
 import { signal } from "@preact/signals";
