+++ frontend/src/form/types/FormRowMode.ts
