+++ frontend/src/form/types/InputProps.ts
@@ -1,4 +1,5 @@
 import { type ReadonlySignal } from "@preact/signals";
+
 import { StaticInputProps } from "./StaticInputProps";
 import { FormRowMode } from "./FormRowMode";
 import { SetValueFunc } from "./SetValueFunc";
