+++ frontend/src/form/types/StaticInputProps.ts
